Refactor database and tests, remove local
Refactored each database table into a separate file. Added separate
tests for each. Also removed local directory which included the database
file. Manager is pulled out to top level and handles database
interactions directly via transactions, which should allow some safety
with handling partially complete transactions during a failure. Overall
better structure and easier to read.
