+++ manager-rs/dauth-service/src/database/flood_vectors.rs
@@ -1,12 +1,14 @@
 use sqlx::sqlite::{SqlitePool, SqliteRow};
+use sqlx::Error as SqlxError;
 use sqlx::{Sqlite, Transaction};
 
 use crate::data::error::DauthError;
 
-/// Creates the auth vector table if it does not exist already.
+/// Creates the flood vector table if it does not exist already.
+/// Meant to be used before the auth vector table.
 pub async fn init_table(pool: &SqlitePool) -> Result<(), DauthError> {
     sqlx::query(
-        "CREATE TABLE IF NOT EXISTS auth_vector_table (
+        "CREATE TABLE IF NOT EXISTS flood_vector_table (
             user_id TEXT NOT NULL,
             seqnum INT NOT NULL,
             xres_star_hash BLOB NOT NULL,
@@ -20,8 +22,6 @@ pub async fn init_table(pool: &SqlitePool) -> Result<(), DauthError> {
     Ok(())
 }
 
-/* Queries */
-
 /// Inserts a vector with the given data.
 /// Returns an error if (id, seqnum) is not unique.
 pub async fn add(
@@ -33,7 +33,7 @@ pub async fn add(
     rand: &[u8],
 ) -> Result<(), DauthError> {
     sqlx::query(
-        "INSERT INTO auth_vector_table
+        "INSERT INTO flood_vector_table
         VALUES ($1,$2,$3,$4,$5)",
     )
     .bind(id)
@@ -47,20 +47,26 @@ pub async fn add(
     Ok(())
 }
 
-/// Returns the first for a given id, sorted by rank (seqnum).
+/// Check first for higher priority flood vectors
+/// Generally, it is normal for this to return Ok(None)
 pub async fn get_first(
     transaction: &mut Transaction<'_, Sqlite>,
     id: &str,
-) -> Result<SqliteRow, DauthError> {
-    Ok(sqlx::query(
-        "SELECT * FROM auth_vector_table
+) -> Result<Option<SqliteRow>, SqlxError> {
+    let res = sqlx::query(
+        "SELECT * FROM flood_vector_table
         WHERE user_id=$1
         ORDER BY seqnum
         LIMIT 1;",
     )
     .bind(id)
     .fetch_one(transaction)
-    .await?)
+    .await;
+
+    match res {
+        Err(SqlxError::RowNotFound) => Ok(None),
+        _ => Ok(Some(res?)),
+    }
 }
 
 /// Removes the vector with the (id, seqnum) pair.
@@ -70,7 +76,7 @@ pub async fn remove(
     seqnum: i64,
 ) -> Result<(), DauthError> {
     sqlx::query(
-        "DELETE FROM auth_vector_table
+        "DELETE FROM flood_vector_table
         WHERE (user_id,seqnum)=($1,$2)",
     )
     .bind(id)
@@ -83,13 +89,13 @@ pub async fn remove(
 /// Removes all vectors belonging to an id.
 pub async fn remove_all(
     transaction: &mut Transaction<'_, Sqlite>,
-    id: &str,
+    user_id: &str,
 ) -> Result<(), DauthError> {
     sqlx::query(
-        "DELETE FROM auth_vector_table
+        "DELETE FROM flood_vector_table
         WHERE user_id=$1",
     )
-    .bind(id)
+    .bind(user_id)
     .execute(transaction)
     .await?;
     Ok(())
@@ -102,11 +108,11 @@ mod tests {
     use rand::distributions::Alphanumeric;
     use rand::{thread_rng, Rng};
     use sqlx::{Row, SqlitePool};
-    use tempfile::tempdir;
+    use tempfile::{tempdir, TempDir};
 
     use auth_vector::constants::{AUTN_LENGTH, RAND_LENGTH, RES_STAR_HASH_LENGTH};
 
-    use crate::local::database::{auth_vectors, general};
+    use crate::database::{flood_vectors, general};
 
     fn gen_name() -> String {
         let s: String = thread_rng().sample_iter(&Alphanumeric).take(10).collect();
@@ -114,15 +120,15 @@ mod tests {
         format!("sqlite_{}.db", s)
     }
 
-    async fn init() -> SqlitePool {
+    async fn init() -> (SqlitePool, TempDir) {
         let dir = tempdir().unwrap();
         let path = String::from(dir.path().join(gen_name()).to_str().unwrap());
         println!("Building temporary db: {}", path);
 
         let pool = general::build_pool(&path).await.unwrap();
-        auth_vectors::init_table(&pool).await.unwrap();
+        flood_vectors::init_table(&pool).await.unwrap();
 
-        pool
+        (pool, dir)
     }
 
     /// Test that db and table creation will work
@@ -133,23 +139,23 @@ mod tests {
 
     /// Test that db fails on non-existent result
     #[tokio::test]
-    #[should_panic]
     async fn test_empty_get() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
-        auth_vectors::get_first(&mut transaction, "test_id_0")
+        assert!(flood_vectors::get_first(&mut transaction, "test_id_0")
             .await
-            .unwrap();
+            .unwrap()
+            .is_none());
 
         transaction.commit().await.unwrap();
     }
 
     /// Test that db can insert
     #[tokio::test]
-    async fn test_get() {
-        let pool = init().await;
+    async fn test_add() {
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
@@ -158,7 +164,7 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                auth_vectors::add(
+                flood_vectors::add(
                     &mut transaction,
                     &format!("test_id_{}", section),
                     row,
@@ -178,10 +184,10 @@ mod tests {
     #[tokio::test]
     #[should_panic]
     async fn test_add_dupicate_fail() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
         let mut transaction = pool.begin().await.unwrap();
 
-        auth_vectors::add(
+        flood_vectors::add(
             &mut transaction,
             "test_id_1",
             1,
@@ -192,7 +198,7 @@ mod tests {
         .await
         .unwrap();
 
-        auth_vectors::add(
+        flood_vectors::add(
             &mut transaction,
             "test_id_1",
             1,
@@ -209,11 +215,11 @@ mod tests {
     /// Test that db can delete after inserts
     #[tokio::test]
     async fn test_get_first() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
-        auth_vectors::add(
+        flood_vectors::add(
             &mut transaction,
             "test_id_1",
             2,
@@ -224,7 +230,7 @@ mod tests {
         .await
         .unwrap();
 
-        auth_vectors::add(
+        flood_vectors::add(
             &mut transaction,
             "test_id_1",
             0,
@@ -235,7 +241,7 @@ mod tests {
         .await
         .unwrap();
 
-        auth_vectors::add(
+        flood_vectors::add(
             &mut transaction,
             "test_id_1",
             1,
@@ -250,11 +256,14 @@ mod tests {
 
         let mut transaction = pool.begin().await.unwrap();
 
-        let res = auth_vectors::get_first(&mut transaction, "test_id_1")
+        let res = flood_vectors::get_first(&mut transaction, "test_id_1")
             .await
-            .unwrap();
+            .unwrap().unwrap();
 
-        assert_eq!("test_id_1", res.get_unchecked::<&str, &str>("user_id"));
+        assert_eq!(
+            "test_id_1",
+            res.get_unchecked::<&str, &str>("user_id")
+        );
         assert_eq!(0, res.get_unchecked::<i64, &str>("seqnum"));
 
         transaction.commit().await.unwrap();
@@ -263,7 +272,7 @@ mod tests {
     /// Test that db can delete after inserts
     #[tokio::test]
     async fn test_remove() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
@@ -272,7 +281,7 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                auth_vectors::add(
+                flood_vectors::add(
                     &mut transaction,
                     &format!("test_id_{}", section),
                     row,
@@ -294,7 +303,7 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                auth_vectors::remove(&mut transaction, &format!("test_id_{}", section), row)
+                flood_vectors::remove(&mut transaction, &format!("test_id_{}", section), row)
                     .await
                     .unwrap();
             }
@@ -306,7 +315,7 @@ mod tests {
     /// Test that db can delete after inserts
     #[tokio::test]
     async fn test_remove_all() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
@@ -315,7 +324,7 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                auth_vectors::add(
+                flood_vectors::add(
                     &mut transaction,
                     &format!("test_id_{}", section),
                     row,
@@ -335,7 +344,7 @@ mod tests {
         let num_sections = 10;
 
         for section in 0..num_sections {
-            auth_vectors::remove_all(&mut transaction, &format!("test_id_{}", section))
+            flood_vectors::remove_all(&mut transaction, &format!("test_id_{}", section))
                 .await
                 .unwrap();
         }
@@ -348,9 +357,9 @@ mod tests {
 
         for section in 0..num_sections {
             assert!(
-                auth_vectors::get_first(&mut transaction, &format!("test_id_{}", section))
+                flood_vectors::get_first(&mut transaction, &format!("test_id_{}", section))
                     .await
-                    .is_err()
+                    .unwrap().is_none()
             )
         }
 
@@ -360,7 +369,7 @@ mod tests {
     /// Test that db can delete after inserts
     #[tokio::test]
     async fn test_get_first_with_delete() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
@@ -369,7 +378,7 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                auth_vectors::add(
+                flood_vectors::add(
                     &mut transaction,
                     &format!("test_id_{}", section),
                     row,
@@ -392,11 +401,11 @@ mod tests {
         for section in 0..num_sections {
             for row in 0..num_rows {
                 let res =
-                    auth_vectors::get_first(&mut transaction, &format!("test_id_{}", section))
+                    flood_vectors::get_first(&mut transaction, &format!("test_id_{}", section))
                         .await
-                        .unwrap();
+                        .unwrap().unwrap();
 
-                auth_vectors::remove(&mut transaction, &format!("test_id_{}", section), row)
+                flood_vectors::remove(&mut transaction, &format!("test_id_{}", section), row)
                     .await
                     .unwrap();
