+++ manager-rs/dauth-service/src/database/kseafs.rs
@@ -6,9 +6,9 @@ use crate::data::error::DauthError;
 /// Creates the kseaf table if it does not exist already.
 pub async fn init_table(pool: &SqlitePool) -> Result<(), DauthError> {
     sqlx::query(
-        "CREATE TABLE IF NOT EXISTS key_share_table (
-            xres_star_hash BLOB PRIMARY KEY,
-            key_share BLOB NOT NULL
+        "CREATE TABLE IF NOT EXISTS kseaf_table (
+            kseaf_uuid BLOB PRIMARY KEY,
+            kseaf_data BLOB NOT NULL
         );",
     )
     .execute(pool)
@@ -18,48 +18,48 @@ pub async fn init_table(pool: &SqlitePool) -> Result<(), DauthError> {
 
 /* Queries */
 
-/// Inserts a key share
+/// Inserts a kseaf with a given uuid and value.
 pub async fn add(
     transaction: &mut Transaction<'_, Sqlite>,
-    xres_star_hash: &[u8],
-    key_share: &[u8],
+    uuid: &[u8],
+    value: &[u8],
 ) -> Result<(), DauthError> {
     sqlx::query(
-        "INSERT INTO key_share_table
+        "INSERT INTO kseaf_table
         VALUES ($1,$2)",
     )
-    .bind(xres_star_hash)
-    .bind(key_share)
+    .bind(uuid)
+    .bind(value)
     .execute(transaction)
     .await?;
 
     Ok(())
 }
 
-/// Returns a key share if found.
+/// Returns a kseaf value if found.
 pub async fn get(
     transaction: &mut Transaction<'_, Sqlite>,
-    xres_star_hash: &[u8],
+    uuid: &[u8],
 ) -> Result<SqliteRow, DauthError> {
     Ok(sqlx::query(
-        "SELECT * FROM key_share_table
-        WHERE xres_star_hash=$1;",
+        "SELECT * FROM kseaf_table
+        WHERE kseaf_uuid=$1;",
     )
-    .bind(xres_star_hash)
+    .bind(uuid)
     .fetch_one(transaction)
     .await?)
 }
 
-/// Deletes a key share if found.
+/// Deletes a kseaf vaule if found.
 pub async fn remove(
     transaction: &mut Transaction<'_, Sqlite>,
-    xres_star_hash: &[u8],
+    uuid: &[u8],
 ) -> Result<(), DauthError> {
     sqlx::query(
-        "DELETE FROM key_share_table
-        WHERE xres_star_hash=$1",
+        "DELETE FROM kseaf_table
+        WHERE kseaf_uuid=$1",
     )
-    .bind(xres_star_hash)
+    .bind(uuid)
     .execute(transaction)
     .await?;
 
@@ -73,11 +73,11 @@ mod tests {
     use rand::distributions::Alphanumeric;
     use rand::{thread_rng, Rng};
     use sqlx::{Row, SqlitePool};
-    use tempfile::tempdir;
+    use tempfile::{tempdir, TempDir};
 
-    use auth_vector::constants::{KSEAF_LENGTH, RES_STAR_HASH_LENGTH, RES_STAR_LENGTH};
+    use auth_vector::constants::{KSEAF_LENGTH, RES_STAR_LENGTH};
 
-    use crate::local::database::{general, key_shares};
+    use crate::database::{general, kseafs};
 
     fn gen_name() -> String {
         let s: String = thread_rng().sample_iter(&Alphanumeric).take(10).collect();
@@ -85,15 +85,15 @@ mod tests {
         format!("sqlite_{}.db", s)
     }
 
-    async fn init() -> SqlitePool {
+    async fn init() -> (SqlitePool, TempDir) {
         let dir = tempdir().unwrap();
         let path = String::from(dir.path().join(gen_name()).to_str().unwrap());
         println!("Building temporary db: {}", path);
 
         let pool = general::build_pool(&path).await.unwrap();
-        key_shares::init_table(&pool).await.unwrap();
+        kseafs::init_table(&pool).await.unwrap();
 
-        pool
+        (pool, dir)
     }
 
     /// Test that db and table creation will work
@@ -105,7 +105,7 @@ mod tests {
     /// Test that insert works
     #[tokio::test]
     async fn test_add() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
@@ -114,9 +114,9 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                key_shares::add(
+                kseafs::add(
                     &mut transaction,
-                    &[section * num_rows + row; RES_STAR_HASH_LENGTH],
+                    &[section * num_rows + row; RES_STAR_LENGTH],
                     &[section * num_rows + row; KSEAF_LENGTH],
                 )
                 .await
@@ -129,7 +129,7 @@ mod tests {
     /// Tests that get works
     #[tokio::test]
     async fn test_get() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
@@ -138,50 +138,7 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                key_shares::add(
-                    &mut transaction,
-                    &[section * num_rows + row; RES_STAR_HASH_LENGTH],
-                    &[section * num_rows + row; KSEAF_LENGTH],
-                )
-                .await
-                .unwrap();
-            }
-        }
-
-        transaction.commit().await.unwrap();
-        let mut transaction = pool.begin().await.unwrap();
-
-        for section in 0..num_sections {
-            for row in 0..num_rows {
-                let res = key_shares::get(
-                    &mut transaction,
-                    &[section * num_rows + row; RES_STAR_HASH_LENGTH],
-                )
-                .await
-                .unwrap();
-
-                assert_eq!(
-                    &[section * num_rows + row; KSEAF_LENGTH],
-                    res.get_unchecked::<&[u8], &str>("key_share")
-                );
-            }
-        }
-        transaction.commit().await.unwrap();
-    }
-
-    /// Test that deletes work
-    #[tokio::test]
-    async fn test_remove() {
-        let pool = init().await;
-
-        let mut transaction = pool.begin().await.unwrap();
-
-        let num_rows = 10;
-        let num_sections = 10;
-
-        for section in 0..num_sections {
-            for row in 0..num_rows {
-                key_shares::add(
+                kseafs::add(
                     &mut transaction,
                     &[section * num_rows + row; RES_STAR_LENGTH],
                     &[section * num_rows + row; KSEAF_LENGTH],
@@ -196,7 +153,50 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                key_shares::remove(
+                let res = kseafs::get(
+                    &mut transaction,
+                    &[section * num_rows + row; RES_STAR_LENGTH],
+                )
+                .await
+                .unwrap();
+
+                assert_eq!(
+                    &[section * num_rows + row; KSEAF_LENGTH],
+                    res.get_unchecked::<&[u8], &str>("kseaf_data")
+                );
+            }
+        }
+        transaction.commit().await.unwrap();
+    }
+
+    /// Test that deletes work
+    #[tokio::test]
+    async fn test_remove() {
+        let (pool, _dir) = init().await;
+
+        let mut transaction = pool.begin().await.unwrap();
+
+        let num_rows = 10;
+        let num_sections = 10;
+
+        for section in 0..num_sections {
+            for row in 0..num_rows {
+                kseafs::add(
+                    &mut transaction,
+                    &[section * num_rows + row; RES_STAR_LENGTH],
+                    &[section * num_rows + row; KSEAF_LENGTH],
+                )
+                .await
+                .unwrap();
+            }
+        }
+
+        transaction.commit().await.unwrap();
+        let mut transaction = pool.begin().await.unwrap();
+
+        for section in 0..num_sections {
+            for row in 0..num_rows {
+                kseafs::remove(
                     &mut transaction,
                     &[section * num_rows + row; RES_STAR_LENGTH],
                 )
@@ -211,7 +211,7 @@ mod tests {
         for section in 0..num_sections {
             for row in 0..num_rows {
                 // should have been deleted
-                assert!(key_shares::get(
+                assert!(kseafs::get(
                     &mut transaction,
                     &[section * num_rows + row; RES_STAR_LENGTH],
                 )
@@ -225,7 +225,7 @@ mod tests {
     /// Test that gets work before a delete
     #[tokio::test]
     async fn test_get_with_delete() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
 
         let mut transaction = pool.begin().await.unwrap();
 
@@ -234,7 +234,7 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                key_shares::add(
+                kseafs::add(
                     &mut transaction,
                     &[section * num_rows + row; RES_STAR_LENGTH],
                     &[section * num_rows + row; KSEAF_LENGTH],
@@ -249,7 +249,7 @@ mod tests {
 
         for section in 0..num_sections {
             for row in 0..num_rows {
-                let res = key_shares::get(
+                let res = kseafs::get(
                     &mut transaction,
                     &[section * num_rows + row; RES_STAR_LENGTH],
                 )
@@ -258,12 +258,12 @@ mod tests {
 
                 assert_eq!(
                     &[section * num_rows + row; KSEAF_LENGTH],
-                    res.get_unchecked::<&[u8], &str>("key_share")
+                    res.get_unchecked::<&[u8], &str>("kseaf_data")
                 );
 
-                key_shares::remove(
+                kseafs::remove(
                     &mut transaction,
-                    res.get_unchecked::<&[u8], &str>("xres_star_hash"),
+                    res.get_unchecked::<&[u8], &str>("kseaf_uuid"),
                 )
                 .await
                 .unwrap();
@@ -276,7 +276,7 @@ mod tests {
         for section in 0..num_sections {
             for row in 0..num_rows {
                 // should have been deleted
-                assert!(key_shares::get(
+                assert!(kseafs::get(
                     &mut transaction,
                     &[section * num_rows + row; RES_STAR_LENGTH],
                 )
@@ -291,10 +291,10 @@ mod tests {
     #[tokio::test]
     #[should_panic]
     async fn test_add_dupicate_fail() {
-        let pool = init().await;
+        let (pool, _dir) = init().await;
         let mut transaction = pool.begin().await.unwrap();
 
-        key_shares::add(
+        kseafs::add(
             &mut transaction,
             &[0_u8; RES_STAR_LENGTH],
             &[1_u8; KSEAF_LENGTH],
@@ -302,7 +302,7 @@ mod tests {
         .await
         .unwrap();
 
-        key_shares::add(
+        kseafs::add(
             &mut transaction,
             &[0_u8; RES_STAR_LENGTH],
             &[1_u8; KSEAF_LENGTH],
