+++ manager-rs/dauth-service/src/database/utilities.rs
