--- manager-rs/dauth-service/src/local/mod.rs
@@ -1,2 +1,0 @@
-pub mod database;
-pub mod manager;
